requirements.txt
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
opencv-python==4.8.0.76
Pillow==9.4.0
torch==2.0.1
pandas==1.5.3
requests==2.31.0
matplotlib==3.7.1
ultralytics==8.0.180
scipy==1.10.1
kiwisolver==1.4.5
fonttools==4.42.1
contourpy==1.1.0
pyparsing==3.1.1
torchvision==0.15.2
PyYAML==6.0.1
tqdm==4.66.1
seaborn==0.12.2
psutil==5.9.5
GitPython==3.1.36